requirements.txt
alembic==1.4.3
aniso8601==8.1.0
attrs==20.3.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
pkg-resources==0.0.0
pycparser==2.20
PyJWT==2.0.0
PyMySQL==0.10.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
shared==1.0.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.22
watchdog==1.0.1
Werkzeug==1.0.1
